\name{arate}
\alias{arate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ calculate MCMC acceptance rate }
\description{
}
\usage{
arate(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ MCMC chain (univariate) }
}
\details{

}
\value{

}
\references{ }
\author{Michael Dietze }
\note{ 
}
\seealso{  }
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }

