% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSpatialCov.matrix.R
\name{calcSpatialCov.matrix}
\alias{calcSpatialCov.matrix}
\title{calcSpatialCov.matrix}
\usage{
\method{calcSpatialCov}{matrix}(x, psi, tau, ...)
}
\arguments{
\item{x}{spatial distance matrix}

\item{psi}{spatial corr}

\item{tau}{spatial var}

\item{...}{additional arguments (currently unused)}
}
\value{
spatial covariance matrix
}
\description{
Currently assumes an exponential spatial dependency
}
\details{
Can make gaussian by passing squared distance matrix
}
\author{
Michael Dietze
}
