% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSpatialCov.R
\name{calcSpatialCov}
\alias{calcSpatialCov}
\title{calcSpatialCov}
\usage{
calcSpatialCov(d, psi, tau, ...)
}
\arguments{
\item{d}{either a spatial distance matrix or a list of component spatial distance matrices}

\item{psi}{spatial corr}

\item{tau}{spatial var}

\item{...}{Additional arguments}
}
\description{
calcSpatialCov
}
\author{
Michael Dietze
}
