% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance12.matrix.R
\name{distance12.martix}
\alias{distance12.martix}
\alias{distance12.matrix}
\title{distance12.matrix}
\usage{
distance12.matrix(x, n1, power = 1)
}
\arguments{
\item{power}{}
}
\value{
d
}
\description{
distance12.matrix
}
\author{
Michael Dietze
}
