% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance12.matrix.R
\name{distance12_matrix}
\alias{distance12_matrix}
\title{distance12_matrix}
\usage{
distance12_matrix(x, n1, power = 1)
}
\arguments{
\item{x}{matrix of locations in parameter space}

\item{n1}{number of rows in the original dataset}

\item{power}{exponent used for calculating distance, default value of 2 = Pythagorean distance}
}
\value{
cross-distance matrix of dimensions n x (n - n1)
}
\description{
distance12_matrix
}
\author{
Michael Dietze
}
