% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{distance_matrix}
\usage{
distance_matrix(x, power = 1, dim = 2)
}
\arguments{
\item{x}{matrix of locations in physical or parameter space}

\item{power}{exponent used for calculating distance, default value of 2 = Pythagorean distance}

\item{dim}{dimension}
}
\value{
symmetric matrix of pairwise squared distances
}
\description{
distance_matrix
}
\author{
Michael Dietze
}
