% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussProcess.R
\name{GaussProcess}
\alias{GaussProcess}
\title{GaussProcess}
\usage{
GaussProcess(x, y, isotropic = TRUE, nugget = TRUE, method = "bayes",
  ngibbs = 5000, burnin = 1000, thin = 1, jump.ic = c(1.1, 0.2),
  prior = "IG", mix = "joint", psi = NULL, zeroMean = FALSE,
  exclude = NULL, ...)
}
\arguments{
\item{nugget}{allows additional error in Y rather than fix interpolation to go through points}

\item{prior}{'unif', 'IG'}

\item{mix}{joint=mix over psi simultanously, each=mix over psi individually}

\item{exclude}{}
}
\description{
GaussProcess
}
\author{
Michael Dietze
}
