% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.GP.R
\name{get_ss}
\alias{get_ss}
\title{get_ss}
\usage{
get_ss(gp, xnew, pos.check)
}
\arguments{
\item{gp}{Gaussian Process}

\item{xnew}{new x coordinate}

\item{pos.check}{check if value needs to be positive (if TRUE, returns -Inf when GP is negative)}
}
\description{
get_ss
}
