% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.GP.R
\name{get_y}
\alias{get_y}
\title{get_y}
\usage{
get_y(SSnew, xnew, llik.fn, priors, llik.par)
}
\arguments{
\item{SSnew}{new summary statistic}

\item{xnew}{new x coordinate}

\item{llik.fn}{list that contains likelihood functions}

\item{priors}{prior list}

\item{llik.par}{parameters to be passed llik functions}
}
\description{
get_y
}
