% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_mle.R
\name{gp_mle}
\alias{gp_mle}
\title{gp_mle}
\usage{
gp_mle(theta, d, nugget, myY, maxval = Inf)
}
\arguments{
\item{d}{spatial distance matrix}

\item{nugget}{allows additional error in Y rather than fix interpolation to go through points}

\item{myY}{}

\item{maxval}{}
}
\value{
val
}
\description{
gp_mle
}
\author{
Michael Dietze
}
