% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_mle.R
\name{gp_mle2}
\alias{gp_mle2}
\title{gp_mle2}
\usage{
gp_mle2(theta, d, nugget, myY, maxval = Inf)
}
\arguments{
\item{d}{spatial distance matrix}

\item{nugget}{allows additional error in Y rather than fix interpolation to go through points}

\item{myY}{}
}
\description{
zero mean version
}
