% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.GP.R
\name{gpeval}
\alias{gpeval}
\title{gpeval}
\usage{
gpeval(xnew, k, mu, tau, psi, x, rng, splinefcns)
}
\arguments{
\item{rng}{range}

\item{splinefcns}{}
}
\description{
Calculates the probability of a set of parameter values, given by xnew
}
\author{
Michael Dietze
}
