% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.GP.R
\name{is.accepted}
\alias{is.accepted}
\title{is.accepted}
\usage{
is.accepted(ycurr, ynew, format = "lin")
}
\arguments{
\item{ycurr}{current value on y axis}

\item{ynew}{new y coordinate}

\item{format}{lin = lnlike fcn, log = log(lnlike)}
}
\description{
is.accepted
}
