% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{jump-class}
\alias{jump-class}
\title{define a class for automatically tuning jump distributions}
\description{
define a class for automatically tuning jump distributions
}
\author{
Michael Dietze
}
