% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump.R
\name{jump}
\alias{jump}
\title{jump}
\usage{
jump(ic = 0, rate = 0.4, ...)
}
\arguments{
\item{ic}{optional data vector}

\item{rate}{target acceptance rate}

\item{...}{Addtional arguments}
}
\description{
jump
}
\author{
Michael Dietze
}
