% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldinvgamma.R
\name{ldinvgamma}
\alias{ldinvgamma}
\title{Log-dinvgamma, based on MCMCpack}
\usage{
ldinvgamma(x, shape, scale = 1)
}
\arguments{
\item{x}{vector of quantiles}

\item{shape, scale}{shape and scale parameters for the inverse Gamma distribution}
}
\value{
log density of the inverse Gamma distribution evaluated at \code{x}
}
\description{
Log-dinvgamma, based on MCMCpack
}
\author{
Michael Dietze
}
