% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.GP.R
\name{minimize.GP}
\alias{minimize.GP}
\title{minimize.GP}
\usage{
minimize.GP(gp, rng, x0, splinefuns = NULL)
}
\arguments{
\item{x0}{}

\item{splinefuns}{}
}
\description{
minimize.GP
}
\author{
Michael Dietze
}
