% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump.R
\name{mvjump}
\alias{mvjump}
\title{mvjump}
\usage{
mvjump(ic = 0, rate = 0.4, nc = 2, ...)
}
\arguments{
\item{ic}{optional data vector}

\item{rate}{target acceptance rate}

\item{nc}{NetCDF object containing target variable}

\item{...}{Additional arguments}
}
\description{
multivariate version
}
