% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.jump.R
\name{p.jump}
\alias{p.jump}
\title{Extract current jump parameter value for a \code{jump} object}
\usage{
\method{p}{jump}(x, ...)
}
\arguments{
\item{x}{object of class \code{jump}}

\item{...}{additional arguments (currently unused)}
}
\value{
The most recent jump parameter value.
}
\description{
Extract current jump parameter value for a \code{jump} object
}
\author{
Michael Dietze
}
