% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.jump.R
\name{p.mvjump}
\alias{p.mvjump}
\title{Extract current jump parameter values for a \code{mvjump} object}
\usage{
\method{p}{mvjump}(x, ...)
}
\arguments{
\item{x}{object of class \code{mvjump}}

\item{...}{additional arguments (currently unused)}
}
\value{
The most recent row of the multivariate jump parameter history.
}
\description{
Extract current jump parameter values for a \code{mvjump} object
}
\author{
Michael Dietze
}
