% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEcAn.emulator-package.R
\docType{package}
\name{PEcAn.emulator-package}
\alias{PEcAn.emulator}
\alias{PEcAn.emulator-package}
\title{Implementation of a Gaussian Process model for `PEcAn`
Supports both likelihood and bayesian approaches for kriging and model
 emulation. Includes functions for sampling design and prediction.}
\description{
Implementation of a Gaussian Process model (both likelihood and bayesian approaches) for kriging and model emulation. Includes functions for sampling design and prediction.
}
\author{
\strong{Maintainer}: Mike Dietze \email{dietze@bu.edu}

Other contributors:
\itemize{
  \item University of Illinois, NCSA [copyright holder]
}

}
\keyword{internal}
