% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jump.R
\name{plot.jump}
\alias{plot.jump}
\title{plot.jump}
\usage{
\method{plot}{jump}(x, ...)
}
\arguments{
\item{x}{object of class \code{jump}}

\item{...}{additional arguments (currently unused)}
}
\value{
Invisible \code{NULL}; called for its side-effect of producing a plot.
}
\description{
Plot diagnostic history for a \code{jump} tuning object.

Displays the jump parameter trace and acceptance rate over MCMC iterations.
}
\author{
Michael Dietze
}
