% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.density.R
\name{predict.density}
\alias{predict.density}
\title{Simple interpolation of a density object to new points}
\usage{
\method{predict}{density}(object, xnew, ...)
}
\arguments{
\item{object}{a \code{density} object (as returned by \code{\link[stats]{density}})}

\item{xnew}{numeric vector of new x coordinates at which to evaluate the density}

\item{...}{additional arguments (currently unused)}
}
\value{
numeric vector of interpolated density values at \code{xnew}
}
\description{
Simple interpolation of a density object to new points
}
\author{
Michael Dietze
}
