% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GP.R
\name{predict.GP}
\alias{predict.GP}
\title{Predict method for Gaussian Process objects}
\usage{
\method{predict}{GP}(object, xpred, cI = NULL, pI = NULL, splinefcns = NULL, ...)
}
\arguments{
\item{object}{a Gaussian Process object (class \code{GP}) as returned by \code{\link{GaussProcess}}}

\item{xpred}{value of x where prediction should be made}

\item{cI}{credible interval quantiles, or \code{NULL} to skip}

\item{pI}{prediction interval quantiles, or \code{NULL} to skip}

\item{splinefcns}{optional list of spline functions for trend-surface adjustment}

\item{...}{additional arguments (currently unused)}
}
\value{
Kriged predictions or a list of credible/prediction interval quantiles.
}
\description{
Performs kriging prediction from a fitted Gaussian Process model.
}
\author{
Michael Dietze
}
