% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.GP.R
\name{summarize_GP}
\alias{summarize_GP}
\title{summarize_GP}
\usage{
summarize_GP(gp, pdf_file = NULL, txt_file = NULL)
}
\arguments{
\item{gp}{Gaussian Process object}

\item{pdf_file}{filename for PDF output of figures (optional)}

\item{txt_file}{filename for text output of summaries (optional)}
}
\description{
Produces diagnostic plots and summary statistics for a fitted Gaussian Process object.
}
\author{
Michael Dietze
}
