% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSpatialCov.list.R
\name{calcSpatialCov.list}
\alias{calcSpatialCov.list}
\title{Currently assumes an exponential spatial dependency}
\usage{
\method{calcSpatialCov}{list}(d, psi, tau, ...)
}
\arguments{
\item{d}{list of component spatial distance matrices}

\item{psi}{spatial corr}

\item{tau}{spatial var}

\item{...}{additional arguments (currently unused)}
}
\value{
spatial covariance matrix
}
\description{
Can make gaussian by passing squared distance matrix
}
\author{
Michael Dietze
}
