% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSpatialCov.matrix.R
\name{calcSpatialCov.matrix}
\alias{calcSpatialCov.matrix}
\title{calcSpatialCov.matrix}
\usage{
\method{calcSpatialCov}{matrix}(d, psi, tau)
}
\arguments{
\item{d}{spatial distance matrix}

\item{psi}{spatial corr}

\item{tau}{spatial var}
}
\description{
Currently assumes an exponential spatial dependency
}
\details{
can make gaussian by passing squared distance matrix
}
\author{
Michael Dietze
}
