% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSpatialCov.R
\name{calcSpatialCov}
\alias{calcSpatialCov}
\title{calcSpatialCov}
\usage{
calcSpatialCov(x, ...)
}
\arguments{
\item{x}{Name of variable to plot on X axis}

\item{...}{Adiitional arguments}
}
\description{
calcSpatialCov
}
\author{
Michael Dietze
}
