% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSpatialCov.R
\name{calcSpatialCov}
\alias{calcSpatialCov}
\title{calcSpatialCov}
\usage{
calcSpatialCov(x, ...)
}
\arguments{
\item{x}{either a spatial distance matrix or a list of component spatial distance matrices}

\item{...}{Additional arguments}
}
\description{
calcSpatialCov
}
\author{
Michael Dietze
}
