% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.matrix.R
\name{distance.martix}
\alias{distance.martix}
\alias{distance.matrix}
\title{distance.matrix}
\usage{
distance.matrix(x, power = 1, dim = 2)
}
\arguments{
\item{dim}{}
}
\value{
d
}
\description{
distance.matrix
}
\author{
Michael Dietze
}
