% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{distance}
\usage{
distance(x, power = 1)
}
\arguments{
\item{x}{Name of variable to plot on X axis}

\item{power}{exponent used for calculating distance, default value of 2 = Pythagorean distance}
}
\value{
dst
}
\description{
distance
}
\author{
Michael Dietze
}
