\name{emulator-package}
\alias{emulator-package}
\alias{emulator}
\docType{package}
\title{
Gausian Process emulator
}
\description{
Implementation of a Gaussian Process model (both likelihood and bayesian approaches) for kriging and model emulation.  Includes functions for sampling design and prediction.
}
\details{
\tabular{ll}{
Package: \tab emulator\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-10-11\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Michael Dietze

Maintainer: Michael Dietze <mdietze@fas.harvard.edu>
}
\references{

}
\keyword{ package }
\seealso{
}
\examples{
}
