% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GaussProcess}
\alias{GaussProcess}
\title{GaussProcess}
\usage{
GaussProcess(x, y, isotropic = TRUE, nugget = TRUE, method = "bayes",
  ngibbs = 5000, burnin = 1000, thin = 1, jump.ic = c(1.1, 0.2),
  prior = "IG", mix = "joint", psi = NULL, zeroMean = FALSE,
  exclude = NULL, ...)
}
\description{

}

