% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_mle.R
\name{gp_mle2}
\alias{gp_mle2}
\title{gp_mle2}
\usage{
gp_mle2(theta, d, nugget, myY, maxval = Inf)
}
\arguments{
\item{theta}{proposed parameter vector: [mu, tauw, tauv, phi1...phiK]}

\item{d}{spatial distance matrix}

\item{nugget}{allows additional error in Y rather than fix interpolation to go through points}

\item{myY}{vector of observed data}

\item{maxval}{maximum value}
}
\description{
zero mean version
}
