% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldinvgamma.R
\name{ldinvgamma}
\alias{ldinvgamma}
\title{ldinvgamma}
\usage{
ldinvgamma(x, shape, scale = 1)
}
\arguments{
\item{scale}{}
}
\value{
log.density
}
\description{
Log-dinvgamma, based on MCMCpack
}
\author{
Michael Dietze
}
