% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc.R
\name{lhc}
\alias{lhc}
\title{lhc}
\usage{
lhc(x, n.samp)
}
\arguments{
\item{x}{<- list (n.dim x 2)}

\item{n.samp}{number of samples}
}
\description{
Latin Hyper Cube
}
\details{
Simple uniform sampling with LHC permutation
}
\author{
Michael Dietze
}
