% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nderiv.R
\name{nderiv}
\alias{nderiv}
\title{nderiv}
\usage{
nderiv(x, y)
}
\arguments{
\item{x}{Name of variable to plot on X axis}

\item{y}{Name of variable to plot on Y axis}
}
\value{
der
}
\description{
nderiv
}
\author{
Michael Dietze
}
