% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jump.R
\name{plot.mvjump}
\alias{plot.mvjump}
\title{Plot diagnostic history for a \code{mvjump} tuning object}
\usage{
\method{plot}{mvjump}(x, ...)
}
\arguments{
\item{x}{object of class \code{mvjump}}

\item{...}{additional arguments (currently unused)}
}
\description{
Displays the multivariate jump parameter trace and acceptance rate over MCMC iterations.
}
\author{
Michael Dietze
}
