% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.density.R
\name{predict.density}
\alias{predict.density}
\title{predict.density}
\usage{
\method{predict}{density}(den, xnew)
}
\arguments{
\item{xnew}{}
}
\value{
ynew
}
\description{
Simple interpolation of a density object to new points
}
\author{
Michael Dietze
}
