% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GP.R
\name{predict.GP}
\alias{predict.GP}
\title{predict.GP}
\usage{
\method{predict}{GP}(gp, xpred, cI = NULL, pI = NULL, splinefcns = NULL)
}
\arguments{
\item{gp}{Gaussian Process}

\item{xpred}{value of x where prediction should be made}

\item{cI}{credible interval}

\item{pI}{prediction interval}

\item{splinefcns}{spline functions}
}
\description{
predict.GP
}
\author{
Michael Dietze
}
