% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GP.R
\name{predict.GP}
\alias{predict.GP}
\title{predict.GP}
\usage{
\method{predict}{GP}(gp, xpred, cI = NULL, pI = NULL,
  splinefcns = NULL)
}
\arguments{
\item{cI}{credible interval}

\item{pI}{prediction interval}
}
\description{
predict.GP
}
\author{
Michael Dietze
}
