% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.GP.R
\name{summarize.GP}
\alias{summarize.GP}
\title{summarize.GP}
\usage{
summarize.GP(gp, pdf_file = NULL, txt_file = NULL)
}
\arguments{
\item{txt_file}{}
}
\description{
summarize.GP
}
\author{
Michael Dietze
}
