% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.jump.R
\name{update.jump}
\alias{update.jump}
\title{update.jump}
\usage{
\method{update}{jump}(jmp, chain)
}
\arguments{
\item{jmp}{jump parameter}

\item{chain}{mcmc chain}
}
\value{
jmp updated jump parameter
}
\description{
update.jump
}
\author{
Michael Dietze
}
