% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.jump.R
\name{update.mvjump}
\alias{update.mvjump}
\title{update.mvjump}
\usage{
\method{update}{mvjump}(jmp, chain)
}
\arguments{
\item{jmp}{jump parameter}

\item{chain}{mcmc chain}
}
\description{
update.mvjump
}
